Support separate bom/pos/fit property setting (breaks many existing rules!)
Options were replaced by properties. Caution: Properties are now
inherited from default choices! This is a design change.

A property can be cleared with '-', or set with '+' in front of the
property code (a lower-case single-character, see below).

The virtual property '!' stands for all 3 properties 'b' (in_bom), 'p'
(in_pos) and 'f' (fitted).

Examples for property settings:
-! -> do not fit, not in bom, not in pos
-f+bp -> do not fit, but put in bom and pos
+b -> in bom (overrides a default choice that sets -b)

Also, there are now implicit defaults for properties (we can do it since
they are bool). That is, if you only have choices which set '-b', you
don't have to specify '+b' as default. The implicit default is always
the opposite of the only polarity used. If both polarities are used in
choices, there is no implicit default polarity and you must define it
yourself.

Sorry for the current lack of documentation.