Multiple values/props within a single choice definition are now allowed and merged into a single assignment
This is a major change how values (and props) are collected and
assigned. Prior to this change, only a single value argument per choice
definition was allowed. To assign a string containing spaces (the
argument separator), the string had to be quoted or the spaces had to be
escaped. Now all value arguments are joined with one space inbetween to
form a single value assignment. Similar rules apply for props, which are
now merged into a single assignment.
Multiple value/prop assignments from different choice definitions are
still and will stay prohibited.